turretCollisions

+++ src/main/java/com/anet/qtr4tdm/common/entities/CannonTurretEntity.java
@@ -55,7 +55,7 @@ public class CannonTurretEntity extends TurretEntity {
     }
 
     @Override
-    public double getMaxEnergy() {
+    public long getMaxEnergy() {
         return 10000;
     }
